pyproject.toml
[project]
name = "intelligent_tutor_crew"
version = "0.1.0"
description = "An agentic AI tutor built using CrewAI and Gemini"
requires-python = ">=3.9"
dependencies = [
    "crewai",
    "google-generativeai",
    "langchain-google-genai",
    "python-dotenv",
    "pydantic",
    "rich"
]

[project.scripts]
run_crew = "intelligent_tutor_crew:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/intelligent_tutor_crew"]

[tool.hatch.build.targets.editable]
dev-mode-dirs = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]